Font factory is now no longer an interface and instead handles a bunch of generic stuff.
