+++ editor/imgui.ini
@@ -130,6 +130,11 @@ Pos=60,60
 Size=1217,576
 Collapsed=0
 
+[Window][Load list..##filebrowser_140699753734064]
+Pos=367,260
+Size=700,450
+Collapsed=0
+
 [Docking][Data]
 DockSpace         ID=0x8B93E3BD Window=0xA787BDB4 Pos=1743,253 Size=1434,951 Split=Y Selected=0x429E880E
   DockNode        ID=0x00000003 Parent=0x8B93E3BD SizeRef=1434,99 Split=X Selected=0xFE740DCC
