+++ editor/src/vulkan_app.h
@@ -5,7 +5,7 @@
 #include "events/event.h"
 
 #include "moth_ui/iimage_factory.h"
-#include "moth_ui/ifont_factory.h"
+#include "moth_ui/font_factory.h"
 #include "moth_ui/irenderer.h"
 
 #include "vulkan/vulkan_context.h"
@@ -61,7 +61,7 @@ namespace backend::vulkan {
         static char const* const PERSISTENCE_FILE;
 
         std::unique_ptr<moth_ui::IImageFactory> m_imageFactory;
-        std::unique_ptr<moth_ui::IFontFactory> m_fontFactory;
+        std::unique_ptr<moth_ui::FontFactory> m_fontFactory;
         std::unique_ptr<moth_ui::IRenderer> m_uiRenderer;
 
         std::unique_ptr<Graphics> m_graphics;
