+++ include/moth_ui/context.h
@@ -1,14 +1,14 @@
 #pragma once
 
 #include "iimage_factory.h"
-#include "ifont_factory.h"
+#include "font_factory.h"
 #include "irenderer.h"
 
 namespace moth_ui {
     class Context {
     public:
         Context(IImageFactory* imageFactory,
-                IFontFactory* fontFactory,
+                FontFactory* fontFactory,
                 IRenderer* renderer);
         ~Context() = default;
 
@@ -26,7 +26,7 @@ namespace moth_ui {
             return *m_imageFactory;
         }
 
-        IFontFactory& GetFontFactory() const {
+        FontFactory& GetFontFactory() const {
             assert(m_fontFactory && "No font factory?");
             return *m_fontFactory;
         }
@@ -40,7 +40,7 @@ namespace moth_ui {
         static std::shared_ptr<Context> s_currentContext;
 
         IImageFactory* m_imageFactory;
-        IFontFactory* m_fontFactory;
+        FontFactory* m_fontFactory;
         IRenderer* m_renderer;
     };
 }
